transcribe_file/requirements.txt
google-cloud-storage>=1.42.3
google-cloud-pubsub>=2.9.0
google-api-core>=2.2.2
google-cloud-speech>=2.11.1
protobuf>=3.19.1